+++ src/autonavlog/domain/__init__.py
@@ -1,6 +1,5 @@
 from .calculation import CalculationOutcome, Issue, SectionResult
 from .project import Airport, NavSection, Project, RouteNode, VisualReference
-from .snapshot import CalculationSnapshot
 from .values import AdoptedValue
 from .weather import ForecastRequirement, WeatherRequest, WeatherResult
 
@@ -8,7 +7,6 @@ __all__ = [
     "AdoptedValue",
     "Airport",
     "CalculationOutcome",
-    "CalculationSnapshot",
     "ForecastRequirement",
     "Issue",
     "NavSection",